Refactored name to `generate_template_files.py`
New name should avoid confusion when working with the actual `Template`
object, and be more clear as to what the output of the script is (a
collection of .template files, not `Template()` objects).
