+++ generate_template_files.py
